+++ .storybook/preview.ts
@@ -77,4 +77,3 @@ const preview: Preview = {
 };
 
 export default preview;
-