+++ src/components/mediasfuComponents/MediasfuConference.tsx
@@ -4154,6 +4154,8 @@ const MediasfuConference: React.FC<MediasfuConferenceOptions> = ({
         const ip = Object.keys(socket)[0];
 
         // Assuming each socket has a disconnect method
+        socket[ip].removeAllListeners?.();
+        socket[ip].removeAllListeners?.();
         socket[ip].disconnect();
       } catch (error) {
         console.log(
@@ -4179,6 +4181,8 @@ const MediasfuConference: React.FC<MediasfuConferenceOptions> = ({
     updateIsShareEventModalVisible(false);
     updateIsConfirmExitModalVisible(false);
     await disconnectAllSockets(consume_sockets.current);
+    consume_sockets.current = [];
+    consume_sockets.current = [];
     await updateStatesToInitialValues();
     updateMeetingProgressTime("00:00:00");
     updateMeetingElapsedTime(0);