+++ src/components/mediasfuComponents/MediasfuGeneric.tsx
@@ -4627,6 +4627,8 @@ const MediasfuGeneric: React.FC<MediasfuGenericOptions> = ({
         const ip = Object.keys(socket)[0];
 
         // Assuming each socket has a disconnect method
+        socket[ip].removeAllListeners?.();
+        socket[ip].removeAllListeners?.();
         socket[ip].disconnect();
       } catch (error) {
         console.log(
@@ -4652,6 +4654,8 @@ const MediasfuGeneric: React.FC<MediasfuGenericOptions> = ({
     updateIsShareEventModalVisible(false);
     updateIsConfirmExitModalVisible(false);
     await disconnectAllSockets(consume_sockets.current);
+    consume_sockets.current = [];
+    consume_sockets.current = [];
     await updateStatesToInitialValues();
     updateMeetingProgressTime("00:00:00");
     updateMeetingElapsedTime(0);