+++ src/components/mediasfuComponents/MediasfuWebinar.tsx
@@ -4152,6 +4152,8 @@ const MediasfuWebinar: React.FC<MediasfuWebinarOptions> = ({
         const ip = Object.keys(socket)[0];
 
         // Assuming each socket has a disconnect method
+        socket[ip].removeAllListeners?.();
+        socket[ip].removeAllListeners?.();
         socket[ip].disconnect();
       } catch (error) {
         console.log(
@@ -4177,6 +4179,8 @@ const MediasfuWebinar: React.FC<MediasfuWebinarOptions> = ({
     updateIsShareEventModalVisible(false);
     updateIsConfirmExitModalVisible(false);
     await disconnectAllSockets(consume_sockets.current);
+    consume_sockets.current = [];
+    consume_sockets.current = [];
     await updateStatesToInitialValues();
     updateMeetingProgressTime("00:00:00");
     updateMeetingElapsedTime(0);